refactor: badges folder to more
