+++ karman_app/lib/pages/more/details/skeleton_page.dart
