+++ karman_app/lib/pages/more/focus_badges_page.dart
