+++ karman_app/lib/pages/more/habit_badges_page.dart
